Move config utility function to own module in class package
