+++ packages/class/src/validate.ts
@@ -53,13 +53,3 @@ export function parse(input: unknown): Config {
   }
   return input;
 }
-
-type RecursivePartial<T> = {
-  [P in keyof T]?: T[P] extends (infer U)[]
-    ? RecursivePartial<U>[]
-    : T[P] extends object | undefined
-      ? RecursivePartial<T[P]>
-      : T[P];
-};
-
-export type PartialConfig = RecursivePartial<Config>;
